# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Qubirex API",
    "version": "4.3.1-phase0.1",
    "description": "Generated from the Express routers by scripts/openapi.js. Do not edit by hand."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "cookieSession": []
    },
    {
      "bearer": []
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "qbx_session",
        "description": "httpOnly SameSite=Strict session cookie. State-changing requests also need X-CSRF-Token."
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Same session token; accepted until the frontend moves to cookies (docs/decisions.md D-002)."
      }
    },
    "parameters": {
      "CsrfToken": {
        "name": "X-CSRF-Token",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "Required when authenticating with the session cookie."
      }
    },
    "responses": {
      "Error": {
        "description": "Error. New routes (/api/employer, /api/verify, /api/practical) answer {error: {code, message}}; legacy routes answer {error: string}.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "oneOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ]
                    }
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/admin/employers": {
      "get": {
        "tags": [
          "admin"
        ],
        "operationId": "get_admin_employers",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/admin/employers/{id}/kyb": {
      "post": {
        "tags": [
          "admin"
        ],
        "operationId": "post_admin_employers_id_kyb",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ]
      }
    },
    "/api/admin/institutions": {
      "get": {
        "tags": [
          "admin"
        ],
        "operationId": "get_admin_institutions",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/admin/mastery-logs/{id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "operationId": "get_admin_mastery_logs_id",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/admin/ontology-review": {
      "get": {
        "tags": [
          "admin"
        ],
        "operationId": "get_admin_ontology_review",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/admin/ontology-review/{id}": {
      "post": {
        "tags": [
          "admin"
        ],
        "operationId": "post_admin_ontology_review_id",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ]
      }
    },
    "/api/admin/quality-report": {
      "get": {
        "tags": [
          "admin"
        ],
        "operationId": "get_admin_quality_report",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/admin/skills": {
      "get": {
        "tags": [
          "admin"
        ],
        "operationId": "get_admin_skills",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/admin/stats": {
      "get": {
        "tags": [
          "admin"
        ],
        "operationId": "get_admin_stats",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  }
}